requirements.txt
<<<<<<< HEAD
﻿aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
backports.tarfile==1.2.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.2
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.16
contourpy==1.3.2
cryptography==46.0.6
cycler==0.12.1
defusedxml==0.7.1
exceptiongroup==1.3.1
fastapi==0.135.3
filelock==3.25.2
filetype==1.2.0
firebase_admin==7.3.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
imageio-ffmpeg==0.6.0
importlib_metadata==9.0.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
keyring==25.7.0
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
more-itertools==11.0.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.1
numpy==1.26.4
onnx==1.21.0
onnxruntime==1.23.2
onnxslim==0.1.90
opencv-python==4.13.0.92
openvino==2024.6.0
openvino-dev==2024.6.0
openvino-telemetry==2025.2.0
packaging==26.0
pi_heif==1.3.0
pillow==12.2.0
pillow-avif-plugin==1.5.5
polars==1.39.3
polars-runtime-32==1.39.3
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pyparsing==3.3.2
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyttsx3==2.99
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.33.1
razorpay==1.4.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.14.0
torch==2.11.0+cu128
torchaudio==2.11.0+cu128
torchvision==0.26.0+cu128
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.33
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.43.0
websockets==16.0
yarl==1.23.0
yt-dlp==2026.3.17
zipp==3.23.0
=======
numpy
pdfplumber
sentence-transformers
aiohttp
python-dotenv
pyttsx3
pytest
pytest-asyncio
fastapi
uvicorn
>>>>>>> 2c7c158ab4b54348e45911533a25b045f3d7342e
